requirements.txt
azure-ai-inference
azure-ai-agents
azure-ai-projects
azure-identity
azure-keyvault-secrets
azure-search-documents
pydantic>=2.0
pygithub
opentelemetry-sdk
opentelemetry-exporter-otlp
python-dotenv
pytest
pytest-asyncio
